@@ -53,67 +56,79 @@
       "eslint --cache --fix"
     ]
   },
-  "dependencies": {
-    "@next/mdx": "^11.1.0",
-    "github-slugger": "^1.3.0",
-    "gray-matter": "^4.0.3",
-    "lodash": "^4.17.21",
-    "marked": "^2.1.3",
-    "mdx-observable": "0.2.0",
-    "next-mdx-remote": "^3.0.4",
-    "prism-react-renderer": "1.2.1",
-    "react-click-away-listener": "2.0.3",
-    "react-copy-to-clipboard": "^5.0.3",
-    "react-instantsearch-dom": "^6.12.1",
-    "react-markdown": "^6.0.3",
-    "react-scroll": "^1.8.3",
-    "rehype-autolink-headings": "^5.1.0",
-    "rehype-slug": "^4.0.1",
-    "remark-external-links": "^8.0.0",
-    "remark-gfm": "^2.0.0"
-  },
-  "devDependencies": {
-    "@babel/preset-env": "^7.15.0",
-    "@babel/preset-react": "^7.14.5",
-    "@commitlint/cli": "^13.1.0",
-    "@commitlint/config-conventional": "^13.1.0",
-    "@mdx-js/runtime": "^1.6.22",
-    "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-commonjs": "^20.0.0",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.0.4",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.2.1",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-jest": "^24.4.0",
-    "eslint-plugin-mdx": "^1.14.1",
-    "eslint-plugin-react": "^7.24.0",
-    "husky": "^7.0.1",
-    "jest": "^27.0.6",
-    "lint-staged": "^11.1.2",
-    "next": "^11.0.0",
-    "nock": "^13.1.1",
-    "node-fetch": "^2.6.1",
-    "pinst": "^2.1.6",
-    "prettier": "^2.3.2",
-    "react": "*",
-    "react-dom": "*",
-    "rollup": "^2.56.2",
-    "stylelint": "^13.13.1",
-    "stylelint-config-standard": "^22.0.0",
-    "watch": "^1.0.2"
-  },
-  "peerDependencies": {
-    "next": ">=11.x",
-    "react": "~17.x",
-    "react-dom": "~17.x"
-  },
   "release": {
     "branches": [
       "main"
     ]
+  },
+  "dependencies": {
+    "@mdx-js/mdx": "^2.1.3",
+    "@next/mdx": "^12.3.1",
+    "github-slugger": "^1.4.0",
+    "gray-matter": "^4.0.3",
+    "lodash": "^4.17.21",
+    "marked": "^4.1.0",
+    "mdast": "^3.0.0",
+    "next-mdx-remote": "^4.1.0",
+    "prism-react-renderer": "1.3.5",
+    "react-click-away-listener": "2.2.2",
+    "react-copy-to-clipboard": "^5.1.0",
+    "react-instantsearch-dom": "^6.33.0",
+    "react-markdown": "^8.0.3",
+    "react-scroll": "^1.8.7",
+    "rehype-autolink-headings": "^6.1.1",
+    "rehype-sanitize": "^5.0.1",
+    "rehype-slug": "^5.0.1",
+    "remark-comment": "^1.0.0",
+    "remark-external-links": "^9.0.1",
+    "remark-gfm": "^3.0.1",
+    "remark-heading-id": "^1.0.0",
+    "remark-inline-links": "^6.0.1",
+    "remark-parse": "^10.0.1",
+    "strip-indent": "^4.0.0",
+    "unist-util-visit": "^4.1.1",
+    "vfile": "^5.3.5"
+  },
+  "devDependencies": {
+    "@babel/preset-env": "^7.19.1",
+    "@babel/preset-react": "^7.18.6",
+    "@commitlint/cli": "^17.1.2",
+    "@commitlint/config-conventional": "^17.1.0",
+    "@mdx-js/runtime": "^1.6.22",
+    "@rollup/plugin-babel": "^5.3.1",
+    "@rollup/plugin-commonjs": "^22.0.2",
+    "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-node-resolve": "^14.1.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@vitejs/plugin-react": "^2.1.0",
+    "esbuild": "^0.15.9",
+    "eslint": "^8.24.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-mdx": "^2.0.5",
+    "eslint-plugin-react": "^7.31.8",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "husky": "^8.0.1",
+    "jsdom": "^20.0.0",
+    "lint-staged": "^13.0.3",
+    "mdx-observable": "^0.2.0",
+    "next": "^12.3.1",
+    "nock": "^13.2.9",
+    "pinst": "^3.0.0",
+    "pkg-dir": "^7.0.0",
+    "prettier": "^2.7.1",
+    "react": "*",
+    "react-dom": "*",
+    "rollup": "^2.79.1",
+    "vitest": "^0.23.4",
+    "watch": "^1.0.2",
+    "whatwg-fetch": "^3.6.2"
+  },
+  "peerDependencies": {
+    "next": ">12",
+    "react": ">18",
+    "react-dom": ">18"
   },
   "publishConfig": {
     "access": "public"
