@@ -1,1 +1,3 @@
-module.exports = require('./dist/serialize')
+import { pageProps, staticPaths } from './dist/serialize.js'
+
+export { pageProps, staticPaths }
