@@ -1,1 +1,1 @@
-module.exports = require('./dist/index')
+export * from './dist/index.js'
