@@ -1,33 +1,31 @@
 /* eslint-disable react/display-name */
 import React from 'react'
 import { MDXRemote } from 'next-mdx-remote'
-import { State, Observe } from 'mdx-observable'
-import { Element } from 'react-scroll'
+// import { State, Observe } from 'mdx-observable'
+import Tabs, { Tab } from './components/Tabs.jsx'
+// import Interpolate from './components/Interpolate.jsx'
 import CodeBlock from './components/CodeBlock.jsx'
 import InlineCode from './components/InlineCode.jsx'
-import Tabs, { Tab } from './components/Tabs.jsx'
-import Interpolate from './components/Interpolate.jsx'
 
-export default function Documentation({ source, theme }) {
-  const components = {
-    Tabs: Tabs,
-    Tab: Tab,
-    Interpolate,
-    State,
-    Observe,
-    Element: ({ name, ...props }) => {
-      return (
-        <Element
-          // remove name from parent div
-          name={props.children[0]?.props?.id === name ? null : name}
-          {...props}
-        />
-      )
+export function Documentation({ source, theme, additionalComponents = {} }) {
+  const components = Object.assign(
+    {
+      Element: ({ name, ...props }) => {
+        return (
+          <div
+            // remove name from parent div
+            name={props.children[0]?.props?.id === name ? null : name}
+            {...props}
+          />
+        )
+      },
+      pre: (props) => <CodeBlock {...props} theme={theme} />,
+      code: (props) => <InlineCode {...props} theme={theme} />,
+      Tabs,
+      Tab
     },
-    pre: (props) => <div {...props} />,
-    code: (props) => <CodeBlock {...props} theme={theme} />,
-    inlineCode: (props) => <InlineCode {...props} theme={theme} />
-  }
+    additionalComponents
+  )
 
   return <MDXRemote {...source} components={components} theme={theme} />
 }
