@@ -1,57 +1,46 @@
 import GithubSlugger from 'github-slugger'
 
-const hasUnclosedSectionTag = (children) => {
-  const amountOpeningTags = children.filter(
-    (c) => c.value && c.value.includes('<Element')
-  ).length
-  const amountClosingTags = children.filter(
-    (c) => c.value && c.value.includes('</Element')
-  ).length
-  return amountOpeningTags > amountClosingTags
-}
-
 const insertSections = (children, slugger, depth = 6) => {
   if (depth === 0) {
-    while (hasUnclosedSectionTag(children)) {
-      children.push({
-        type: 'jsx',
-        value: `</Element>`
-      })
-    }
-    return children
+    return
   }
-  const newChildren = children.reduce((arr, child) => {
-    const childrenToAdd = []
+  const headingIndex = children.findIndex(
+    (c) => c.type === 'heading' && c.depth === depth
+  )
+  const nextHeadingIndex = children
+    .slice(headingIndex + 1, children.length + 1)
+    .findIndex((c) => c.type === 'heading' && c.depth <= depth)
 
-    if (child.type === 'heading' && child.depth <= depth) {
-      const needsClosingSection = hasUnclosedSectionTag(arr)
+  if (headingIndex === -1) {
+    return insertSections(children, slugger, depth - 1)
+  }
 
-      if (needsClosingSection) {
-        childrenToAdd.push({
-          type: 'jsx',
-          value: `</Element>`
-        })
+  const hasNoNextHeading = nextHeadingIndex === -1
+  const end = hasNoNextHeading ? children.length : nextHeadingIndex + 1
+
+  const sectionChildren = children.slice(headingIndex, headingIndex + end)
+  children.splice(headingIndex, end, {
+    type: 'mdxJsxFlowElement',
+    name: 'Element',
+    attributes: [
+      {
+        type: 'mdxJsxAttribute',
+        name: 'name',
+        value: slugger.slug(children[headingIndex].children[0].value)
       }
+    ],
+    children: sectionChildren,
+    data: {
+      _mdxExplicitJsx: true
     }
-    if (child.type === 'heading' && child.depth === depth) {
-      childrenToAdd.push({
-        type: 'jsx',
-        value: `<Element name="${slugger.slug(child.children[0].value)}">`,
-        depth: child.depth
-      })
-    }
+  })
 
-    childrenToAdd.push(child)
-
-    return [...arr, ...childrenToAdd]
-  }, [])
-  return insertSections(newChildren, slugger, depth - 1)
+  return insertSections(children, slugger)
 }
 
 export default function remarkTabs() {
   return ({ children }) => {
     const slugger = new GithubSlugger()
-    const newChildren = insertSections(children, slugger, 2)
-    children.splice(0, children.length, ...newChildren)
+    insertSections(children, slugger)
   }
 }
