@@ -1,14 +1,19 @@
 import { serialize } from 'next-mdx-remote/serialize'
 import matter from 'gray-matter'
-import marked from 'marked'
+import { marked } from 'marked'
 import GithubSlugger from 'github-slugger'
 import rehypeSlug from 'rehype-slug'
 import rehypeAutolink from 'rehype-autolink-headings'
 import cloneDeep from 'lodash/cloneDeep'
 
+import { codeImport as remarkCodeImport } from './lib/remark-plugins/import'
+import remarkInlineLinks from 'remark-inline-links'
+import remarkHeadingId from './lib/remark-plugins/heading-ids'
+// TODO: plugins require some refactoring, see https://github.com/storybookjs/storybook/issues/9602 for inspiration/guidance
 import remarkTabs from './lib/remark-plugins/tabs'
-import remarkState from './lib/remark-plugins/state'
-import remarkSections from './lib/remark-plugins/sections'
+// import remarkState from './lib/remark-plugins/state'
+// import remarkSections from './lib/remark-plugins/sections'
+import remarkGfm from 'remark-gfm'
 import remarkExternalLinks from 'remark-external-links'
 import remarkInternalLinks from './lib/remark-plugins/links'
 import remarkRewriteImages from './lib/remark-plugins/images'
@@ -19,30 +24,65 @@ import {
   getPaths,
   fetchDocsManifest
 } from './lib/docs'
-import { getRawFileFromRepo } from './lib/github'
+import { getRawFile } from './lib/files'
 
-const DOCS_FOLDER = process.env.DOCS_FOLDER
+const defaults = {
+  docsFolder: 'docs',
+  rootPath: 'content',
+  skipPathPrefix: false,
+  trailingSlash: false,
+  useMDX: false,
+  org: null,
+  repo: null,
+  tag: null,
+  assetsDestination: null,
+  debug: process.env.DEBUG === true
+}
 
-export async function pageProps({ params }) {
+export async function pageProps(context, args) {
+  const options = { ...defaults, ...args }
+  const params = context.params
+  const { docsFolder, trailingSlash, skipPathPrefix, useMDX } = options
+
   const slugger = new GithubSlugger()
-  const manifest = await fetchDocsManifest().catch((error) => {
-    if (error.status === 404) return
-    throw error
-  })
-  const { slug } = getSlug(params)
-  const route = manifest && findRouteByPath(slug, manifest.routes)
-  const manifestRoutes = cloneDeep(manifest.routes)
-  replaceDefaultPath(manifestRoutes)
+  const manifest = await fetchDocsManifest(docsFolder, options).catch(
+    (error) => {
+      if (error.status === 404) return
+      throw error
+    }
+  )
 
+  const { slug } = getSlug(params)
+  const pathPrefix = docsFolder && !skipPathPrefix ? `/${docsFolder}` : ''
+  const route =
+    manifest && findRouteByPath(pathPrefix + slug, manifest.routes, options)
   if (!route)
     return {
       notFound: true
     }
+  const inlineLinkSlugHelper =
+    params.slug?.length > 0 && route.path.endsWith('README.md')
+      ? params.slug?.concat(['README'])
+      : params.slug
+  const manifestRoutes = cloneDeep(manifest.routes)
+  replaceDefaultPath(manifestRoutes, options)
 
-  const mdxRawContent = await getRawFileFromRepo(route.path)
+  const mdxRawContent = await getRawFile(route.path, options)
   const { content, data } = matter(mdxRawContent)
+
+  if (!data.title) {
+    data.title = ''
+  }
+
+  const importBasePath = `${process.cwd()}/content${slug
+    .split('/')
+    .slice(0, -1)
+    .join('/')}`
+
   const mdxSource = await serialize(content, {
+    parseFrontmatter: false,
     scope: { data },
+    format: useMDX === 'true' ? 'mdx' : 'md',
     mdxOptions: {
       rehypePlugins: [
         rehypeSlug,
@@ -57,7 +97,7 @@ export async function pageProps({ params }) {
               type: 'element',
               tagName: 'svg',
               properties: {
-                className: ['h-6', 'w-6', 'ml-2', 'text-pink-600'],
+                className: ['h-6', 'w-6', 'ml-2', 'docs-copy-btn'],
                 xmlns: 'http://www.w3.org/2000/svg',
                 fill: 'none',
                 viewBox: '0 0 24 24',
@@ -81,22 +121,33 @@ export async function pageProps({ params }) {
         ]
       ],
       remarkPlugins: [
-        remarkSections,
+        remarkHeadingId,
+        // remarkSections,
         remarkTabs,
-        remarkState,
-        [remarkRewriteImages, { destination: process.env.ASSETS_DESTINATION }],
+        // remarkState,
+        remarkGfm,
+        [
+          remarkCodeImport,
+          {
+            disabled: options.repo ? true : false, // only works with local filesystem, not remote fetch
+            importBasePath: importBasePath
+          }
+        ],
+        [remarkRewriteImages, { destination: options.assetsDestination }],
+        remarkInlineLinks,
         [remarkExternalLinks, { target: false, rel: ['nofollow'] }],
         [
           remarkInternalLinks,
           {
-            prefix: DOCS_FOLDER,
-            slug: params.slug,
-            extensions: ['.mdx', '.md']
+            prefix: docsFolder,
+            slug: inlineLinkSlugHelper,
+            extensions: ['.mdx', '.md'],
+            trailingSlash: trailingSlash
           }
         ]
       ]
     },
-    target: ['es2020']
+    target: ['esnext']
   })
 
   const markdownTokens = marked.lexer(content)
@@ -106,7 +157,10 @@ export async function pageProps({ params }) {
       heading.slug = slugger.slug(heading.text)
       return heading
     })
-  const title = headings[0].text
+
+  const firstHeadingText =
+    headings && headings.length > 0 ? headings[0].text : ''
+  const title = data.title.length > 0 ? data.title : firstHeadingText
 
   return {
     title,
@@ -118,10 +172,26 @@ export async function pageProps({ params }) {
   }
 }
 
-export async function staticPaths() {
-  const manifest = await fetchDocsManifest()
-  const paths = getPaths(manifest.routes)
-  paths.shift() // remove "/docs/README"
-  paths.unshift(`/${process.env.DOCS_FOLDER}`)
-  return paths
+export async function staticPaths(args) {
+  const options = { ...args, ...defaults }
+  const { docsFolder, skipPathPrefix } = options
+
+  const manifest = await fetchDocsManifest(docsFolder, options)
+  const paths = getPaths(manifest.routes, options).filter(
+    (p) => !p.includes('README')
+  )
+
+  if (skipPathPrefix) {
+    paths.unshift('/')
+  }
+
+  return paths.map((p) => {
+    const parts = p.split('/')
+    parts.shift()
+    return {
+      params: {
+        slug: parts.length > 0 ? parts : null
+      }
+    }
+  })
 }
